# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Nutterx Technologies VCF Registration API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RegistrationInput {
  /** @minLength 1 */
  name: string;
  /**
     * E.164 format phone number
     * @pattern ^\+[1-9]\d{6,14}$
     */
  phone: string;
}

export interface Registration {
  id: number;
  name: string;
  phone: string;
  createdAt: string;
}

export interface RegistrationsPage {
  data: Registration[];
  total: number;
  page: number;
  limit: number;
}

export interface Settings {
  id: number;
  companyName: string;
  title: string;
  phone: string;
  email: string;
  website: string;
  address: string;
  updatedAt: string;
}

export interface SettingsInput {
  companyName: string;
  title: string;
  phone: string;
  email: string;
  website: string;
  address: string;
}

export interface AdminLoginInput {
  username: string;
  password: string;
}

export interface AdminUser {
  username: string;
}

export interface MessageResponse {
  message: string;
}

export interface ErrorResponse {
  error: string;
  message: string;
}

export type GetRegistrationsParams = {
/**
 * Search by name or phone
 */
search?: string;
page?: number;
limit?: number;
};