# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Nutterx Technologies VCF Registration API
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
  - name: registration
    description: Registration operations
  - name: admin
    description: Admin operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"

  /register:
    post:
      operationId: submitRegistration
      tags: [registration]
      summary: Register a phone number
      description: Register a name and phone number in E.164 format
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegistrationInput"
      responses:
        "201":
          description: Registration successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Registration"
        "409":
          description: Phone number already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "429":
          description: Too many requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /download-vcf:
    get:
      operationId: downloadVcf
      tags: [registration]
      summary: Download VCF contact card
      description: Download the Nutterx Technologies VCF contact file
      responses:
        "200":
          description: VCF file download
          content:
            text/vcard:
              schema:
                type: string

  /admin/login:
    post:
      operationId: adminLogin
      tags: [admin]
      summary: Admin login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminLoginInput"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/logout:
    post:
      operationId: adminLogout
      tags: [admin]
      summary: Admin logout
      responses:
        "200":
          description: Logged out
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"

  /admin/me:
    get:
      operationId: getAdminMe
      tags: [admin]
      summary: Get current admin session
      responses:
        "200":
          description: Current admin
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AdminUser"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/registrations:
    get:
      operationId: getRegistrations
      tags: [admin]
      summary: List all registrations
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
          description: Search by name or phone
        - name: page
          in: query
          required: false
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            default: 20
      responses:
        "200":
          description: List of registrations
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RegistrationsPage"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/registrations/export:
    get:
      operationId: exportRegistrationsCsv
      tags: [admin]
      summary: Export registrations as CSV
      responses:
        "200":
          description: CSV file download
          content:
            text/csv:
              schema:
                type: string
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/registrations/{id}:
    delete:
      operationId: deleteRegistration
      tags: [admin]
      summary: Delete a registration
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /admin/settings:
    get:
      operationId: getSettings
      tags: [admin]
      summary: Get company settings for VCF
      responses:
        "200":
          description: Current settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    put:
      operationId: updateSettings
      tags: [admin]
      summary: Update company settings for VCF
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SettingsInput"
      responses:
        "200":
          description: Updated settings
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Settings"
        "401":
          description: Not authenticated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status

    RegistrationInput:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        phone:
          type: string
          pattern: "^\\+[1-9]\\d{6,14}$"
          description: E.164 format phone number
      required:
        - name
        - phone

    Registration:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        - id
        - name
        - phone
        - createdAt

    RegistrationsPage:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Registration"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
      required:
        - data
        - total
        - page
        - limit

    Settings:
      type: object
      properties:
        id:
          type: integer
        companyName:
          type: string
        title:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        address:
          type: string
        updatedAt:
          type: string
          format: date-time
      required:
        - id
        - companyName
        - title
        - phone
        - email
        - website
        - address
        - updatedAt

    SettingsInput:
      type: object
      properties:
        companyName:
          type: string
        title:
          type: string
        phone:
          type: string
        email:
          type: string
        website:
          type: string
        address:
          type: string
      required:
        - companyName
        - title
        - phone
        - email
        - website
        - address

    AdminLoginInput:
      type: object
      properties:
        username:
          type: string
        password:
          type: string
      required:
        - username
        - password

    AdminUser:
      type: object
      properties:
        username:
          type: string
      required:
        - username

    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
      required:
        - error
        - message