# lib/api-zod/src/generated/types/messageResponse.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Nutterx Technologies VCF Registration API
 * OpenAPI spec version: 0.1.0
 */

export interface MessageResponse {
  message: string;
}