# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Nutterx Technologies VCF Registration API
 * OpenAPI spec version: 0.1.0
 */

export interface RegistrationInput {
  /** @minLength 1 */
  name: string;
  /**
     * E.164 format phone number
     * @pattern ^\+[1-9]\d{6,14}$
     */
  phone: string;
}