# lib/api-zod/src/generated/types/registrationsPage.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Nutterx Technologies VCF Registration API
 * OpenAPI spec version: 0.1.0
 */
import type { Registration } from './registration';

export interface RegistrationsPage {
  data: Registration[];
  total: number;
  page: number;
  limit: number;
}