# lib/api-zod/src/generated/types/settingsInput.ts
/**
 * Generated by orval v8.20.0 🍺
 * Do not edit manually.
 * Api
 * Nutterx Technologies VCF Registration API
 * OpenAPI spec version: 0.1.0
 */

export interface SettingsInput {
  companyName: string;
  title: string;
  phone: string;
  email: string;
  website: string;
  address: string;
}